pyproject.toml
[tool.poetry]
name = "asf-v2"
version = "0.1.0"
description = ""
authors = ["anthonymanciny <141266961+anthonymanciny@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
sqlalchemy = "^2.0.36"
pymysql = "^1.1.1"
psycopg2-binary = "^2.9.10"
alembic = "^1.14.0"
python-decouple = "^3.8"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.17"
python-dotenv = "^1.0.1"
mariadb = "^1.1.11"
bcrypt = "^4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
